feat: implement dark mode toggle (#1058)
